+++ changelog.md
@@ -26,6 +26,7 @@
 
 * `[gui][desktop]` ScreenLock logic unified with simLock's
 * `[messages]` Changed fonts of message snippet and its prefix.
+* `[cellular]` Improve MMI call forwarding support
 
 ## [0.47.1 2020-11-20]